requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-docx==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0
docxtpl==0.16.7

